refactor(variant-editor): Move download into utils.
