+++ lib/components/variant-editor.js
@@ -1,10 +1,10 @@
-import debounce from 'debounce'
+import debounce from 'lodash.debounce'
 import React, {Component, PropTypes} from 'react'
 
 import {Button, Group as ButtonGroup} from './buttons'
 import Icon from './icon'
 import {Text} from './input'
-import convertToR5Modification from '../utils/convert-to-r5-modification'
+import download from '../utils/download-variant'
 
 export default class VariantEditor extends Component {
   static propTypes = {
@@ -23,51 +23,19 @@ export default class VariantEditor extends Component {
 
   _export = (variantIndex) => {
     const {feeds, modifications, scenarioName, variants} = this.props
-
-    let r5modifications
-    try {
-      r5modifications = modifications
-        .filter((mod) => mod.variants[variantIndex])
-        .map((mod) => convertToR5Modification(mod))
-    } catch (e) {
-      window.alert(`Can not export variant:\n${e.message}`)
-      return
-    }
-
-    const feedChecksums = {}
-    feeds.forEach((f) => {
-      feedChecksums[f.id] = f.checksum
+    download({
+      description: `${scenarioName}-${variants[variantIndex]}`,
+      feeds,
+      modifications: modifications.filter((m) => m.variants[variantIndex])
     })
-
-    const scenario = {
-      id: 0,
-      // scenario name should probably not be at top level of props, but it is
-      description: scenarioName,
-      feedChecksums,
-      modifications: r5modifications
-    }
-
-    const filename = `${scenarioName}_${variants[variantIndex]}.json`
-      .replace(/[^a-zA-Z0-9]/, '_')
-
-    // pretty print the json
-    const out = JSON.stringify(scenario, null, 2)
-    const uri = `data:application/json;base64,${window.btoa(out)}`
-    const a = document.createElement('a')
-    a.setAttribute('href', uri)
-    a.setAttribute('target', '_blank')
-    a.setAttribute('download', filename)
-    a.click()
   }
 
   _create = () => {
     const {currentScenario, saveScenario} = this.props
     // once upon a time a bug could cause variants to be null. allow recovering from this scenario.
-    const variants = currentScenario.variants ? [...currentScenario.variants] : []
-    variants.push('New variant')
     saveScenario({
       ...currentScenario,
-      variants
+      variants: [...(currentScenario.variants || []), 'New variant']
     })
   }
 
@@ -122,6 +90,13 @@ export default class VariantEditor extends Component {
 }
 
 class Variant extends Component {
+  static propTypes = {
+    index: PropTypes.number.isRequired,
+    name: PropTypes.string.isRequired,
+    projectId: PropTypes.string.isRequired,
+    scenarioId: PropTypes.string.isRequired
+  }
+
   _export = () => {
     const {exportVariant, index} = this.props
     exportVariant(index)
