+++ package.json
@@ -46,6 +46,7 @@
     "isomorphic-fetch": "^2.2.1",
     "leaflet": "^0.7.7",
     "leaflet-draw": "^0.3.0",
+    "lodash.debounce": "^4.0.8",
     "lodash.isequal": "^4.3.1",
     "lodash.omit": "^4.5.0",
     "lonlng": "0.1.0",
